requirements.txt
scrapy==1.1
awesome-slugify==1.6
pytz==2016.4
boto==2.40.0
six==1.10.0
